requirements.txt
pyspark
apache-sedona
pandas

# testing
virtualenv==20.0.33
pre-commit
isort
black
flake8
pytest
